[releng] Move SequenceDDiagramSpec in a subpackage
Move this internal SequenceDDiagram class into internal.spec package.
This is done to help extension of Sirius to identify this specific
package.
