+++ plugins/org.eclipse.sirius.diagram.sequence.model/META-INF/MANIFEST.MF
@@ -14,10 +14,10 @@ Export-Package: org.eclipse.sirius.diagram.sequence;version="7.3.0",
  org.eclipse.sirius.diagram.sequence.description.tool.util;version="7.3.0",
  org.eclipse.sirius.diagram.sequence.description.util;version="7.3.0",
  org.eclipse.sirius.diagram.sequence.impl;version="7.3.0",
- org.eclipse.sirius.diagram.sequence.model.business.internal;version="7.3.0";x-internal:=true,
  org.eclipse.sirius.diagram.sequence.model.business.internal.color;version="7.3.0";x-internal:=true,
  org.eclipse.sirius.diagram.sequence.model.business.internal.description;version="7.3.0";x-internal:=true,
  org.eclipse.sirius.diagram.sequence.model.business.internal.ordering;version="7.3.0";x-internal:=true,
+ org.eclipse.sirius.diagram.sequence.model.business.internal.spec;version="7.3.0";x-internal:=true,
  org.eclipse.sirius.diagram.sequence.model.internal;version="7.3.0";x-internal:=true,
  org.eclipse.sirius.diagram.sequence.ordering;version="7.3.0",
  org.eclipse.sirius.diagram.sequence.ordering.impl;version="7.3.0";x-internal:=true,
