+++ plugins/org.eclipse.sirius.diagram.sequence.model/src/org/eclipse/sirius/diagram/sequence/model/business/internal/spec/SequenceDDiagramSpec.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2010, 2021 THALES GLOBAL SERVICES and others.
+ * Copyright (c) 2010, 2023 THALES GLOBAL SERVICES and others.
  * This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
@@ -10,7 +10,7 @@
  * Contributors:
  *    Obeo - initial API and implementation
  *******************************************************************************/
-package org.eclipse.sirius.diagram.sequence.model.business.internal;
+package org.eclipse.sirius.diagram.sequence.model.business.internal.spec;
 
 import java.util.Collection;
